Ensure deposit due notifications include artist avatar
