+++ backend/tests/test_notifications.py
@@ -975,6 +975,69 @@ def test_deposit_due_notification_includes_artist_business_name():
     app.dependency_overrides.clear()
 
 
+def test_deposit_due_notification_includes_artist_avatar():
+    Session = setup_app()
+    db = Session()
+    client = User(
+        email="client4@test.com",
+        password="x",
+        first_name="C4",
+        last_name="User",
+        user_type=UserType.CLIENT,
+    )
+    artist = User(
+        email="artist4@test.com",
+        password="x",
+        first_name="A4",
+        last_name="Artist",
+        user_type=UserType.ARTIST,
+    )
+    db.add_all([client, artist])
+    db.commit()
+    db.refresh(client)
+    db.refresh(artist)
+    profile = models.ArtistProfile(
+        user_id=artist.id,
+        profile_picture_url="http://example.com/pic.jpg",
+    )
+    db.add(profile)
+    db.commit()
+    db.refresh(profile)
+
+    booking = models.BookingSimple(
+        quote_id=3,
+        artist_id=artist.id,
+        client_id=client.id,
+        payment_status="pending",
+        deposit_amount=75,
+        deposit_due_by=datetime(2026, 1, 1),
+        deposit_paid=False,
+    )
+    db.add(booking)
+    db.commit()
+    db.refresh(booking)
+
+    notify_deposit_due(
+        db,
+        client,
+        booking.id,
+        float(booking.deposit_amount),
+        booking.deposit_due_by,
+    )
+    db.close()
+
+    token = create_access_token({"sub": client.email})
+    client_api = TestClient(app)
+    res = client_api.get(
+        "/api/v1/notifications",
+        headers={"Authorization": f"Bearer {token}"},
+    )
+    assert res.status_code == 200
+    data = res.json()
+    assert data[0]["avatar_url"] == "http://example.com/pic.jpg"
+    app.dependency_overrides.clear()
+
+
 def test_booking_request_api_parses_sender_and_type():
     Session = setup_app()
     db = Session()
